+++ packages/components/src/components/carousel/resources.ts
@@ -1,4 +1,4 @@
-import { IconName } from "../icon/interfaces";
+import { IconName } from "../icon/types";
 
 export const DURATION = 6000;